requirements.txt
torch==2.4.0
torchvision==0.19.0
absl-py==2.1.0
accelerate==1.2.1
aenum==3.1.15
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
bitblas==0.1.0
blis==0.7.11
boto3==1.35.78
botocore==1.35.78
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.1.0
colorama==0.4.6
compressed-tensors==0.6.0
confection==0.1.5
contourpy==1.3.1
cpplint==2.0.0
cycler==0.12.1
cymem==2.0.10
Cython==3.0.11
dacite==1.8.1
DataProperty==1.0.1
datasets==3.2.0
datatrove==0.3.0
deepspeed==0.15.4
dill==0.3.6
diskcache==5.6.3
distro==1.9.0
docker-pycreds==0.4.0
docutils==0.21.2
dtlib==0.0.0.dev2
einops==0.8.0
execnet==2.1.1
fastapi==0.115.6
filelock==3.16.1
flash-attn==2.7.0.post2
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.9.0
fuzzywuzzy==0.18.0
gguf==0.10.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.68.1
h11==0.14.0
hjson==3.1.0
hqq==0.2.2
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.28.1
humanize==4.11.0
idna==3.10
iniconfig==2.0.0
interegular==0.3.3
jieba==0.42.1
Jinja2==3.1.4
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
lark==1.2.2
lighteval @ git+https://github.com/huggingface/lighteval@ea46419a93fb390e8f694f7c6c64c1e684487c9d
llvmlite==0.43.0
lm-format-enforcer==0.10.6
loguru==0.7.3
lxml==5.3.0
marisa-trie==1.2.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.10.0
mbstrdecoder==1.1.3
mistral_common==1.5.1
ml_dtypes==0.5.0
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.1.0
multiprocess==0.70.14
murmurhash==1.0.11
-e git+https://github.com/huggingface/nanotron@2cde8f63519f15aa449803441ec70225644bd25d#egg=nanotron
networkx==3.4.2
ninja==1.11.1.3
nltk==3.9.1
numba==0.60.0
numpy==1.26.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
openai==1.57.4
opencv-python-headless==4.10.0.84
outlines==0.0.46
pandas==2.2.3
partial-json-parser==0.2.1.1.post4
pathvalidate==3.2.1
pillow==10.4.0
pluggy==1.5.0
portalocker==3.0.0
preshed==3.0.9
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.1
propcache==0.2.1
protobuf==3.20.3
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==18.1.0
pyarrow-hotfix==0.6
pycountry==24.6.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pyparsing==3.2.0
pytablewriter==1.2.0
pytest==8.3.4
pytest-xdist==3.6.1
python-dotenv==1.0.1
pytorch-triton==0.0.1
pytz==2024.2
PyYAML==6.0.2
quanto==0.2.0
RapidFuzz==3.11.0
ray==2.40.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rouge==1.0.1
rouge_score==0.1.2
rpds-py==0.22.3
s3transfer==0.10.4
sacrebleu==2.4.3
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==2.19.2
setproctitle==1.3.4
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.0
starlette==0.41.3
sympy==1.13.1
tabledata==1.3.3
tabulate==0.9.0
tcolorpy==0.1.6
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.3.0
thefuzz==0.22.1
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.3
tqdm==4.67.1
transformers==4.46.3
triton==3.0.0
typepy==1.3.2
typer==0.9.4
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
vllm==0.6.3.post1
wandb==0.19.0
wasabi==1.1.3
watchfiles==1.0.3
weasel==0.3.4
websockets==14.1
Werkzeug==3.1.3
word2number==1.1
xformers==0.0.27.post2
xxhash==3.5.0
yarl==1.18.3
